+++ index.js
@@ -10,7 +10,6 @@ module.exports = function(opts) {
 module.exports.meshers = {
   culled: require('./meshers/culled').mesher,
   greedy: require('./meshers/greedy').mesher,
-  greedier: require('./meshers/greedier').mesher,
   monotone: require('./meshers/monotone').mesher,
   stupid: require('./meshers/stupid').mesher
 }
